requirements.txt
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
